HW-26-Django-Tg-Catalog/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
APScheduler==3.6.3
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
cachetools==4.2.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
Django==4.2.1
django-dotenv==1.4.2
django-extensions==3.2.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.5.0
pycparser==2.21
python-telegram-bot==13.14
pytz==2023.3
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tornado==6.1
tzlocal==5.0.1
urllib3==2.0.3
yarl==1.9.2
